Relocate PDF inspection into publisher
