+++ workers/proof/tsconfig.json
@@ -20,7 +20,6 @@
     "./core.ts",
     "./dispatch.ts",
     "./index.ts",
-    "./pdf.ts",
     "./queue.ts"
   ]
 }
