+++ workers/publisher/pdf-inspection.ts
